Renamed tests to pass integration tests
